+++ src/Models/StopPlace.php
@@ -2,11 +2,12 @@
 
 namespace TromsFylkestrafikk\Netex\Models;
 
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use TromsFylkestrafikk\Netex\Scopes\ValidDateScope;
 
 /**
- * TromsFylkestrafikk\Netex\Models\StopPlace
+ * \TromsFylkestrafikk\Netex\Models\StopPlace
  *
  * @property string $id
  * @property string $version
@@ -20,6 +21,7 @@ use TromsFylkestrafikk\Netex\Scopes\ValidDateScope;
  * @property string|null $validToDate
  * @property string|null $topographicPlaceRef
  * @property string|null $parentSiteRef
+ * @property int $active
  * @property \Illuminate\Support\Carbon|null $created_at
  * @property \Illuminate\Support\Carbon|null $updated_at
  * @property-read \Illuminate\Database\Eloquent\Collection|\TromsFylkestrafikk\Netex\Models\StopPlaceAltId[] $altIds
@@ -29,30 +31,32 @@ use TromsFylkestrafikk\Netex\Scopes\ValidDateScope;
  * @property-read \Illuminate\Database\Eloquent\Collection|\TromsFylkestrafikk\Netex\Models\StopQuay[] $quays
  * @property-read int|null $quays_count
  * @property-read \TromsFylkestrafikk\Netex\Models\TopographicPlace|null $topographicPlace
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace query()
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereChanged($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereCreated($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereCreatedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereLatitude($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereLongitude($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereName($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereParentSiteRef($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereStopPlaceType($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereTopographicPlaceRef($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereUpdatedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereValidFromDate($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereValidToDate($value)
- * @method static \Illuminate\Database\Eloquent\Builder|StopPlace whereVersion($value)
+ * @method static Builder|StopPlace active()
+ * @method static Builder|StopPlace newModelQuery()
+ * @method static Builder|StopPlace newQuery()
+ * @method static Builder|StopPlace query()
+ * @method static Builder|StopPlace whereActive($value)
+ * @method static Builder|StopPlace whereChanged($value)
+ * @method static Builder|StopPlace whereCreated($value)
+ * @method static Builder|StopPlace whereCreatedAt($value)
+ * @method static Builder|StopPlace whereId($value)
+ * @method static Builder|StopPlace whereLatitude($value)
+ * @method static Builder|StopPlace whereLongitude($value)
+ * @method static Builder|StopPlace whereName($value)
+ * @method static Builder|StopPlace whereParentSiteRef($value)
+ * @method static Builder|StopPlace whereStopPlaceType($value)
+ * @method static Builder|StopPlace whereTopographicPlaceRef($value)
+ * @method static Builder|StopPlace whereUpdatedAt($value)
+ * @method static Builder|StopPlace whereValidFromDate($value)
+ * @method static Builder|StopPlace whereValidToDate($value)
+ * @method static Builder|StopPlace whereVersion($value)
  * @mixin \Eloquent
  */
 class StopPlace extends Model
 {
+    public $incrementing = false;
     protected $table = 'netex_stop_place';
     protected $keyType = 'string';
-    public $incrementing = false;
 
     protected static function boot()
     {
@@ -78,5 +82,10 @@ class StopPlace extends Model
     public function groups()
     {
         return $this->belongsToMany(GroupOfStopPlaces::class, 'netex_stop_place_group_member');
+    }
+
+    public function scopeActive(Builder $query)
+    {
+        $query->where('active', true);
     }
 }